Move macro_rules! macros into their own file.
